+++ lib/rundock/scenario.rb
@@ -11,5 +11,10 @@ module Rundock
     def run
       @nodes.each(&:run)
     end
+
+    def complete
+      @nodes.each { |n| n.complete(self) }
+      self
+    end
   end
 end